Move common code from YouTubePlayer to SkyTubeApp
